Main hearts play functionality finished, gitignore
Finished main play functionality of `hearts.py`. Game now transitions
through states, asks for plays, and asks for required information for
moves. Hearts moves are still mostly stubbed out in `hearts.py`, but
they're there. Remaining on the todo is pretty much finishing the moves
and their interaction with the game state as well as prettifying the
game.